% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arcsine_CI_1x2.R
\name{Arcsine_CI_1x2}
\alias{Arcsine_CI_1x2}
\title{Arcsine confidence interval}
\usage{
Arcsine_CI_1x2(X, n = NULL, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\description{
The Arcsine confidence interval for the binomial probability
(with Anscombe variance stabilizing transformation)
Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"
}
\examples{
load_chapter(2)
Arcsine_CI_1x2(500)
Arcsine_CI_1x2(100, 5e3, .1)
}
\references{
Anscombe FJ (1948) The transformation of Poisson, binomial and
negative binomial data. Biometrika; 35:246-254
}
