% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bentur_2009}
\alias{bentur_2009}
\title{Airway hyper-responsiveness before and after stem cell transplantation}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
bentur_2009
}
\description{
Airway hyper-responsiveness before and after stem cell transplantation
}
\references{
Bentur et al. (2009)
}
\keyword{datasets}
