% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blaker_exact_CI_1x2.R
\name{Blaker_exact_CI_1x2}
\alias{Blaker_exact_CI_1x2}
\title{The Blaker exact confidence interval}
\usage{
Blaker_exact_CI_1x2(X, n = NULL, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\description{
The Blaker exact confidence interval for the binomial
probability
Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"
}
\examples{
load_chapter(2)
Blaker_exact_CI_1x2(1)
Blaker_exact_CI_1x2(1, 100)
}
\references{
Blaker H (2000) Confidence curves and improved exact
confidence intervals for discrete distributions. The Canadian Journal of
Statistics; 28:783-798
}
