% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blaker_exact_test_1x2.R
\name{Blaker_exact_test_1x2}
\alias{Blaker_exact_test_1x2}
\title{The Blaker exact test}
\usage{
Blaker_exact_test_1x2(X, n, pi0, printresults = TRUE)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{pi0}{a given probability}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\description{
The Blaker exact test for the binomial probability (pi)
H_0: pi = pi0  vs  H_A: pi ~= pi0 (two-sided)
Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"
}
\examples{
load_chapter(2)
Blaker_exact_test_1x2(1)
Blaker_exact_test_1x2(1, 10, .5)
}
\references{
Blaker H (2000) Confidence curves and improved exact
confidence intervals for discrete distributions. The Canadian Journal of
Statistics; 28:783-798
}
