% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreslowDay_homogeneity_test_stratified_2x2.R
\name{BreslowDay_homogeneity_test_stratified_2x2}
\alias{BreslowDay_homogeneity_test_stratified_2x2}
\title{The Breslow-Day test of homogeneity of odds ratios over strata}
\usage{
BreslowDay_homogeneity_test_stratified_2x2(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A list containing lower bound, upper bound and differences of the statistic
}
\description{
The Breslow-Day test of homogeneity of odds ratios over strata with

Tarone correction

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\examples{
BreslowDay_homogeneity_test_stratified_2x2(doll_hill_1950)
BreslowDay_homogeneity_test_stratified_2x2(hine_1989)
}
