% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calc_prob}
\alias{calc_prob}
\title{Calculate probability}
\usage{
calc_prob(...)
}
\arguments{
\item{...}{arguments passed to methods}
}
\description{
Calculate probability
}
\note{
This function has little use to the user, it is exported for confirmity
to R package standards.
}
