% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calculate_limit_lower}
\alias{calculate_limit_lower}
\title{Calculate the lower limit of a confidence interval}
\usage{
calculate_limit_lower(...)
}
\arguments{
\item{...}{arguments passed to methods}
}
\description{
Calculate the lower limit of a confidence interval
}
\note{
This function has little use to the user, it is exported so that
it can be used by \code{\link[stats:uniroot]{stats::uniroot()}}.
}
