% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-ch2.R
\docType{data}
\name{ch2}
\alias{ch2}
\title{The Agresti-Coull confidence interval for the binomial probability}
\format{
An object of class \code{module} (inherits from \code{list}) of length 2.
}
\usage{
ch2
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (\code{FALSE} = no, \code{TRUE} = yes)}
}
\description{
Described in Chapter 2 "The 1x2 Table and the Binomial
Distribution"

Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"
}
\references{
Agresti A, Coull BA (1998) Approximate is better than "exact"
for interval estimation of binomial proportions. The American
Statistician; 52:119-126
}
\seealso{
Wald_CI_1x2
}
\keyword{datasets}
