% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chacko_test_1xc.R
\name{Chacko_test_1xc}
\alias{Chacko_test_1xc}
\title{The Chacko test for order-restriction}
\usage{
Chacko_test_1xc(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (a 1xc vector, where c is the number of categories)}

\item{printresults}{display results (F = no, T = yes)}
}
\value{
A data frame containing the two-sided p-value, the statistic and the degrees of freedom
}
\description{
The Chacko test for order-restriction

Described in Chapter 3 "The 1xc Table and the Multinomial Distribution"
}
\examples{
# Hypothetical experiment
Chacko_test_1xc(n=c(1, 4, 3, 11, 9))
}
