% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap1}
\alias{chap1}
\title{Chapter 1: Introduction}
\description{
There are no functions for Chapter 1 (Introduction), only from
Chapters 2 to 10.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
