% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap10}
\alias{chap10}
\title{Chapter 10: Stratified 2x2 Tables and Meta-Analysis}
\description{
These are the functions related to chapter 10:
\enumerate{
\item{BreslowDay_homogeneity_test_stratified_2x2}
\item{CochranMantelHaenszel_test_stratified_2x2}
\item{Cochran_Q_test_stratified_2x2}
\item{InverseVariance_estimate_stratified_2x2}
\item{ML_estimates_and_CIs_stratified_2x2}
\item{MantelHaenszel_estimate_stratified_2x2}
\item{Pearson_LR_homogeneity_test_stratified_2x2}
\item{Pearson_LR_test_common_effect_stratified_2x2}
\item{Peto_homogeneity_test_stratified_2x2}
\item{Peto_OR_estimate_stratified_2x2}
\item{RBG_test_and_CI_stratified_2x2}
\item{Wald_test_and_CI_common_diff_stratified_2x2}
\item{Wald_test_and_CI_common_ratio_stratified_2x2}
\item{Woolf_test_and_CI_stratified_2x2}
\item{stratified_2x2_tables}
}
}
\note{
You can also print the list above with \code{list_functions(10)}.
}
\section{Table orientation}{
 In most study designs, rows designate a grouping of the data, for
instance, into treatment or exposure groups, while the columns indicate a
dichotomous outcome, usually with the first column representing the outcome
of interest. This designation of rows and columns may not be relevant in all
study designs, please see the introduction to chapter 10 for details.
}

\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
