% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap2}
\alias{chap2}
\title{Chapter 2: The 1x2 Table and the Binomial Distribution}
\description{
These are the functions related to chapter 2:
\enumerate{
\item{AgrestiCoull_CI_1x2}
\item{Arcsine_CI_1x2}
\item{Wald_CI_1x2}
\item{Blaker_exact_CI_1x2}
\item{Blaker_exact_test_1x2}
\item{Blaker_midP_CI_1x2}
\item{Blaker_midP_test_1x2}
\item{ClopperPearson_exact_CI_1x2}
\item{ClopperPearson_midP_CI_1x2}
\item{Exact_binomial_test_1x2}
\item{Jeffreys_CI_1x2}
\item{LR_CI_1x2}
\item{LR_test_1x2}
\item{MidP_binomial_test_1x2}
\item{Score_test_1x2}
\item{Score_test_CC_1x2}
\item{Wald_CI_CC_1x2}
\item{Wilson_score_CI_1x2}
\item{Wilson_score_CI_CC_1x2}
\item{the_1x2_table_CIs}
\item{Wald_test_1x2}
\item{Wald_test_CC_1x2}
\item{the_1x2_table_tests}
}
}
\note{
You can also print the list above with \code{list_functions(2)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
