% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap3}
\alias{chap3}
\title{Chapter 3: The 1xc Table and the Multinomial Distribution}
\description{
These are the functions related to chapter 3:
\enumerate{
\item{Chacko_test_1xc}
\item{Exact_multinomial_test_1xc}
\item{Gold_Wald_CIs_1xc}
\item{Goodman_Wald_CIs_1xc}
\item{Goodman_Wald_CIs_for_diffs_1xc}
\item{Goodman_Wilson_score_CIs_1xc}
\item{LR_test_1xc}
\item{MidP_multinomial_test_1xc}
\item{Pearson_chi_squared_test_1xc}
\item{QuesenberryHurst_Wilson_score_CIs_1xc}
\item{the_1xc_table_CIs}
\item{the_1xc_table_tests}
}
}
\note{
You can also print the list above with \code{list_functions(3)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
