% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap5}
\alias{chap5}
\title{Chapter 5: The Ordered rx2 Table}
\description{
These are the functions related to chapter 5:
\enumerate{
\item{CochranArmitage_CI_rx2}
\item{CochranArmitage_exact_cond_midP_tests_rx2}
\item{CochranArmitage_MH_tests_rx2}
\item{Exact_cond_midP_unspecific_ordering_rx2}
\item{Pearson_LR_tests_unspecific_ordering_rx2}
\item{the_rx2_table}
\item{Trend_estimate_CI_tests_rx2}
}
}
\note{
You can also print the list above with \code{list_functions(5)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
