% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap6}
\alias{chap6}
\title{Chapter 6: The Ordered 2xc Table}
\description{
These are the functions related to chapter 6:
\enumerate{
\item{Brant_test_2xc}
\item{Cumulative_models_for_2xc}
\item{Exact_cond_midP_linear_rank_tests_2xc}
\item{ClopperPearson_exact_CI_1x2_beta_version}
\item{Exact_cond_midP_unspecific_ordering_rx2}
\item{MantelHaenszel_test_2xc}
\item{Pearson_LR_tests_cum_OR_2xc}
\item{Score_test_for_effect_in_the_probit_model_2xc}
\item{the_2xc_table}
}
}
\note{
You can also print the list above with \code{list_functions(6)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
