% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap7}
\alias{chap7}
\title{Chapter 7: The rxc Table}
\description{
These are the functions related to chapter 7:
\enumerate{
\item{Bonferroni_type_CIs_rxc}
\item{Cumulative_models_for_rxc}
\item{Exact_cond_midP_tests_rxc}
\item{FisherFreemanHalton_asymptotic_test_rxc}
\item{gamma_coefficient_rxc_bca}
\item{gamma_coefficient_rxc}
\item{JonckheereTerpstra_test_rxc}
\item{Kendalls_tau_b_rxc}
\item{Kendalls_tau_b_rxc_bca}
\item{KruskalWallis_asymptotic_test_rxc}
\item{linear_by_linear_test_rxc}
\item{Pearson_correlation_coefficient_rxc}
\item{Pearson_correlation_coefficient_rxc_bca}
\item{Pearson_LR_tests_rxc}
\item{Pearson_residuals_rxc}
\item{Scheffe_type_CIs_rxc}
\item{Spearman_correlation_coefficient_rxc}
\item{Spearman_correlation_coefficient_rxc_bca}
\item{the_rxc_table}
}
}
\note{
You can also print the list above with \code{list_functions(7)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
