% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap9}
\alias{chap9}
\title{Chapter 9: The Paired cxc Table}
\description{
These are the functions related to chapter 9:
\enumerate{
\item{Bhapkar_test_paired_cxc}
\item{Bonferroni_type_CIs_paired_cxc}
\item{FleissEveritt_test_paired_cxc}
\item{FleissLevinPaik_test_paired_cxc}
\item{McNemarBowker_test_paired_cxc}
\item{Scheffe_type_CIs_paired_cxc}
\item{Score_test_and_CI_marginal_mean_scores_paired_cxc}
\item{Stuart_test_paired_cxc}
\item{Wald_test_and_CI_marginal_mean_ranks_paired_cxc}
\item{Wald_test_and_CI_marginal_mean_scores_paired_cxc}
\item{the_paired_cxc_table_nominal}
\item{the_paired_cxc_table_ordinal}
}
}
\note{
You can also print the list above with \code{list_functions(9)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
