% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranArmitage_exact_cond_midP_tests_rx2.R
\name{CochranArmitage_exact_cond_midP_tests_rx2}
\alias{CochranArmitage_exact_cond_midP_tests_rx2}
\title{The Cochran-Armitage exact conditional and mid-P tests}
\usage{
CochranArmitage_exact_cond_midP_tests_rx2(n, a, printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (an rx2 matrix)}

\item{a}{scores assigned to the rows}

\item{printresults}{display results}
}
\value{
A data frame containing the two-sided, twice-the-smallest tail P-value and the mid-P value
}
\description{
The Cochran-Armitage exact conditional and mid-P tests

Described in Chapter 5 "The Ordered rx2 Table"
}
\examples{
\dontrun{
CochranArmitage_exact_cond_midP_tests_rx2(mills_graubard_1987, c(1, 2, 3, 4, 5))
}
CochranArmitage_exact_cond_midP_tests_rx2(indredavik_2008, c(1, 2, 3, 4, 5))
}
