% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranArmitage_MH_tests_rx2.R
\name{CochranArmitage_MH_tests_rx2}
\alias{CochranArmitage_MH_tests_rx2}
\title{The Cochran-Armitage, modified Cochran-Armitage, and Mantel-Haenszel tests for trend}
\usage{
CochranArmitage_MH_tests_rx2(n, a)
}
\arguments{
\item{n}{the observed counts (an rx2 matrix)}

\item{a}{scores assigned to the rows}
}
\value{
A list containing observed statistics and p-values
}
\description{
Described in Chapter 5 "The Ordered rx2 Table"
}
\examples{
CochranArmitage_MH_tests_rx2(mills_graubard_1987, c(1, 2, 3, 4, 5))
CochranArmitage_MH_tests_rx2(indredavik_2008, c(1, 2, 3, 4, 5))
}
