% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencytables-package.R
\docType{package}
\name{contingencytables}
\alias{contingencytables-package}
\alias{contingencytables}
\title{Statistical Analysis of Contingency tables}
\description{
Statistical Analysis of Contingency Tables is an invaluable tool for statistical inference in contingency tables. It covers effect size estimation, confidence intervals, and hypothesis tests for the binomial and the multinomial distributions, unpaired and paired 2x2 tables, rxc tables, ordered rx2 and 2xc tables, paired cxc tables, and stratified tables.
This package provides functions that accompany the "Statistical Analysis of Contingency Tables" book by Fagerland et. al. <ISBN 9781466588172>.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
\item https://ocbe-uio.github.io/contingencytables/
}
}
\seealso{
\link{print.contingencytables_result} to read about printing alternatives.
}
\author{
\strong{Maintainer}: Waldir Leoncio \email{w.l.netto@medisin.uio.no}

Authors:
\itemize{
  \item Morten Wang Fagerland \email{morten.fagerland@medisin.uio.no}
}

Other contributors:
\itemize{
  \item Stian Lydersen [contributor]
  \item Petter Laake [contributor]
  \item Ole Christian Lingjærde [translator]
  \item Brad J. Biggerstaff [contributor]
}

}
