% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulative_models_for_2xc.R
\name{Cumulative_models_for_2xc}
\alias{Cumulative_models_for_2xc}
\title{Cumulative logit and probit models}
\usage{
Cumulative_models_for_2xc(
  n,
  linkfunction = "logit",
  alpha = 0.05,
  printresults = TRUE
)
}
\arguments{
\item{n}{the observed table (a 2xc matrix)}

\item{linkfunction}{either "logit" or "probit"}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\value{
A list containing the results of statistical tests for the
goodness-of-fit of a proportional odds model, the effect in a proportional
odds model and the effect parameter beta in the proportional odds model.
}
\description{
Cumulative logit and probit models

Described in Chapter 6 "The Ordered 2xc Table"
}
\examples{
Cumulative_models_for_2xc(fontanella_2008)
Cumulative_models_for_2xc(lydersen_2012a)
}
