% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulative_models_for_rxc.R
\name{Cumulative_models_for_rxc}
\alias{Cumulative_models_for_rxc}
\title{Cumulative logit and probit models}
\usage{
Cumulative_models_for_rxc(
  n,
  linkfunction = "logit",
  alpha = 0.05,
  printresults = TRUE
)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{linkfunction}{either "logit" or "probit"}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A list containing the results of statistical tests for the
goodness-of-fit of a proportional odds model, the effect in a proportional
odds model and row comparisons
}
\description{
Cumulative logit and probit models

Described in Chapter 7 "The rxc Table"
}
\examples{
Cumulative_models_for_rxc(table_7.5)
Cumulative_models_for_rxc(table_7.6)
}
