% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.onAttach}
\alias{.onAttach}
\title{Prints welcome message on package load}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{no idea, but will break devtools::document() if removed}

\item{pkgname}{name of the package}
}
\description{
Prints package version number and welcome message on package load
}
