% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.3}
\alias{table_7.3}
\title{Treatment for ear infection}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 2 columns.
}
\usage{
table_7.3
}
\description{
Treatment for ear infection
}
\keyword{datasets}
