% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencytables-class.R
\name{fetch_print_format}
\alias{fetch_print_format}
\title{Fetch print format}
\usage{
fetch_print_format(x)
}
\arguments{
\item{x}{a named vector, data.frame or list containing test statistics}
}
\value{
A string with the proper print_structure for \code{x}
printing formats for the contingencytables_result class
}
\description{
This is just a way of automating the retrieval of common
}
