% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherFreemanHalton_asymptotic_test_rxc.R
\name{FisherFreemanHalton_asymptotic_test_rxc}
\alias{FisherFreemanHalton_asymptotic_test_rxc}
\title{The Fisher-Freeman-Halton asymptotic test for unordered rxc tables}
\usage{
FisherFreemanHalton_asymptotic_test_rxc(n)
}
\arguments{
\item{n}{the observed counts (an rxc matrix)}
}
\value{
A list containing the probability, the statistic and the degrees of freedom
}
\description{
The Fisher-Freeman-Halton asymptotic test for unordered rxc tables

Described in Chapter 7 "The rxc Table"
}
\note{
May not give results for all tables, due to overflow
}
\examples{
FisherFreemanHalton_asymptotic_test_rxc(table_7.3)
}
