% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleissEveritt_test_paired_cxc.R
\name{FleissEveritt_test_paired_cxc}
\alias{FleissEveritt_test_paired_cxc}
\title{The Fleiss-Everitt version of the Stuart test for marginal homogeneity}
\usage{
FleissEveritt_test_paired_cxc(n)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}
}
\value{
A list containing the probability, the statistic and the degrees of freedom
}
\description{
The Fleiss-Everitt version of the Stuart test for marginal homogeneity

Described in Chapter 9 "The Paired cxc Table"
}
\examples{
FleissEveritt_test_paired_cxc(fleiss_2003)
}
