% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fontanella_2008}
\alias{fontanella_2008}
\title{The Adolescent Placement Study}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 4 columns.
}
\usage{
fontanella_2008
}
\description{
The Adolescent Placement Study
}
\references{
Fontanella et al. (2008)
}
\keyword{datasets}
