% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_coefficient_rxc_bca.R
\name{gamma_coefficient_rxc_bca}
\alias{gamma_coefficient_rxc_bca}
\title{The gamma coefficient with the bias-corrected and accelerated boostrap confidence interval}
\usage{
gamma_coefficient_rxc_bca(n, nboot = 10000, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{nboot}{number of bootstrap samples}

\item{alpha}{the nominal significance level, used to compute a 100(1-alpha) confidence interval}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
a list with the gamma coefficient and the confidence interval limits
}
\description{
The gamma coefficient with the bias-corrected and accelerated boostrap confidence interval

Described in Chapter 7 "The rxc Table"
}
\examples{
\dontrun{
  gamma_coefficient_rxc_bca(table_7.7)
  gamma_coefficient_rxc_bca(table_7.8)
  gamma_coefficient_rxc_bca(table_7.9)
}
}
