% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hine_1989}
\alias{hine_1989}
\title{Prophylactice use of Lidocaine in myocardial infarction}
\format{
An object of class \code{array} of dimension 2 x 2 x 6.
}
\usage{
hine_1989
}
\description{
Prophylactice use of Lidocaine in myocardial infarction
}
\references{
Hine et al. (1989)
}
\keyword{datasets}
