% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hypothetical}
\alias{hypothetical}
\title{Hypothetical experiment}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
hypothetical
}
\description{
Hypothetical experiment
}
\keyword{datasets}
