% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{indredavik_2008}
\alias{indredavik_2008}
\title{Elevated troponin T levels in stroke patients}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 2 columns.
}
\usage{
indredavik_2008
}
\description{
Elevated troponin T levels in stroke patients
}
\references{
Indredavik et al. (2008)
}
\keyword{datasets}
