% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InverseVariance_estimate_stratified_2x2.R
\name{InverseVariance_estimate_stratified_2x2}
\alias{InverseVariance_estimate_stratified_2x2}
\title{The inverse variance estimate of the overall effect across strata}
\usage{
InverseVariance_estimate_stratified_2x2(n, link = "logit")
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}

\item{link}{the link function ('linear', 'log', or 'logit')}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The inverse variance estimate of the overall effect across strata

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\examples{
InverseVariance_estimate_stratified_2x2(doll_hill_1950)
InverseVariance_estimate_stratified_2x2(hine_1989)
}
