% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jeffreys_CI_1x2.R
\name{Jeffreys_CI_1x2}
\alias{Jeffreys_CI_1x2}
\title{Jeffreys confidence interval for the binomial probability}
\usage{
Jeffreys_CI_1x2(X, n, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (0 = no, 1 <- yes)}
}
\value{
A vector containing lower, upper and point estimates of the statistic
}
\description{
Jeffreys confidence interval for the binomial probability

Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"
}
\examples{

# The number of 1st order male births (Singh et al. 2010)
Jeffreys_CI_1x2(singh_2010["1st", "X"], singh_2010["1st", "n"])
# The number of 2nd order male births (Singh et al. 2010)
Jeffreys_CI_1x2(singh_2010["2nd", "X"], singh_2010["2nd", "n"])
# The number of 3rd order male births (Singh et al. 2010)
Jeffreys_CI_1x2(singh_2010["3rd", "X"], singh_2010["3rd", "n"])
# The number of 4th order male births (Singh et al. 2010)
with(singh_2010["4th", ], Jeffreys_CI_1x2(X, n)) # alternative syntax
# Ligarden et al. (2010)
Jeffreys_CI_1x2(ligarden_2010["X"], ligarden_2010["n"])
}
