% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JonckheereTerpstra_test_rxc.R
\name{JonckheereTerpstra_test_rxc}
\alias{JonckheereTerpstra_test_rxc}
\title{The Jonckheere-Terpstra test for association}
\usage{
JonckheereTerpstra_test_rxc(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\value{
a list containing the standard normalized Jonckheere-Terpstra test statistic
}
\description{
The Jonckheere-Terpstra test for association

Described in Chapter 7 "The rxc Table"
}
\examples{
JonckheereTerpstra_test_rxc(table_7.7)
JonckheereTerpstra_test_rxc(table_7.8)
JonckheereTerpstra_test_rxc(table_7.9)
}
