% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kendalls_tau_b_rxc_bca.R
\name{Kendalls_tau_b_rxc_bca}
\alias{Kendalls_tau_b_rxc_bca}
\title{Kendall's tau-b with the bias-corrected and accelerated boostrap confidence interval}
\usage{
Kendalls_tau_b_rxc_bca(n, nboot = 10000, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{nboot}{number of bootstrap samples}

\item{alpha}{the nominal significance level, used to compute a 100(1-alpha) confidence interval}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\value{
A list containing the statistic and the confindence interval limits
}
\description{
Kendall's tau-b with the bias-corrected and accelerated boostrap confidence interval

Described in Chapter 7 "The rxc Table"
}
\examples{
\dontrun{
  Kendalls_tau_b_rxc_bca(table_7.7)
  Kendalls_tau_b_rxc_bca(table_7.8)
  Kendalls_tau_b_rxc_bca(table_7.9)
}
}
