% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KruskalWallis_asymptotic_test_rxc.R
\name{KruskalWallis_asymptotic_test_rxc}
\alias{KruskalWallis_asymptotic_test_rxc}
\title{The Kruskal-Wallis asymptotic test for singly ordered rxc tables}
\usage{
KruskalWallis_asymptotic_test_rxc(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (an rxc matrix)}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\value{
A list containing the two-sided p-value, the statistic and the degrees of freedom
}
\description{
The Kruskal-Wallis asymptotic test for singly ordered rxc tables

Described in Chapter 7 "The rxc Table"
}
\examples{
KruskalWallis_asymptotic_test_rxc(table_7.5)
KruskalWallis_asymptotic_test_rxc(table_7.6)
}
