% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ligarden_2010}
\alias{ligarden_2010}
\title{Ligarden et al., 2010}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
ligarden_2010
}
\description{
Ligarden et al., 2010
}
\references{
ligarden_2010
}
\keyword{datasets}
