% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{list_functions}
\alias{list_functions}
\title{List functions from a chapter}
\usage{
list_functions(chap_num)
}
\arguments{
\item{chap_num}{Number of book chapter (from 2 to 10)}
}
\value{
List of functions from that chapter
}
\description{
Complements the \code{?chapX} command by printing a list of
functions related to a particular chapter \code{X} on the R console.
}
\author{
Waldir Leoncio
}
