% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{load_chapter}
\alias{load_chapter}
\title{Load functions from a chapter}
\usage{
load_chapter(chap_num)
}
\arguments{
\item{chap_num}{Number of book chapter (from 2 to 10)}
}
\value{
Loads the functions from that chapter
}
\description{
Loads all the functions associated with a specific book chapter.
}
\references{
https://stackoverflow.com/q/64388629/1169233
}
\author{
Waldir Leoncio
}
