% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test_1x2.R
\name{LR_test_1x2}
\alias{LR_test_1x2}
\title{The likelihood ratio test for the binomial probability (pi)}
\usage{
LR_test_1x2(X, n, pi0, printresults = TRUE)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{pi0}{a given probability}

\item{printresults}{display results (0 = no, 1 = yes)}
}
\value{
A vector containing the two-sided p-value, the statistic and the degrees of freedom
}
\description{
The likelihood ratio test for the binomial probability (pi)
H_0: pi = pi0  vs  H_A: pi ~= pi0 (two-sided).
Described in Chapter 2 "The 1x2 Table and the Binomial Distribution".
}
\examples{
LR_test_1x2(singh_2010["1st", "X"], singh_2010["1st", "n"], pi0 = .5)
LR_test_1x2(singh_2010["2nd", "X"], singh_2010["2nd", "n"], pi0 = .5)
LR_test_1x2(singh_2010["3rd", "X"], singh_2010["3rd", "n"], pi0 = .5)
LR_test_1x2(singh_2010["4th", "X"], singh_2010["4th", "n"], pi0 = .5)
LR_test_1x2(ligarden_2010["X"], ligarden_2010["n"], pi0 = .5)
}
