% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lydersen_2012a}
\alias{lydersen_2012a}
\title{Postoperative nausea}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 4 columns.
}
\usage{
lydersen_2012a
}
\description{
Postoperative nausea
}
\references{
Lydersen et al. (2012a)
}
\keyword{datasets}
