% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MantelHaenszel_estimate_stratified_2x2.R
\name{MantelHaenszel_estimate_stratified_2x2}
\alias{MantelHaenszel_estimate_stratified_2x2}
\title{The Mantel-Haenszel estimate of the overall effect across strata}
\usage{
MantelHaenszel_estimate_stratified_2x2(n, link = "logit")
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}

\item{link}{the link function ('linear', 'log', or 'logit')}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Mantel-Haenszel estimate of the overall effect across strata

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\section{Table orientation}{
 In most study designs, rows designate a grouping of the data, for
instance, into treatment or exposure groups, while the columns indicate a
dichotomous outcome, usually with the first column representing the outcome
of interest. This designation of rows and columns may not be relevant in all
study designs, please see the introduction to chapter 4 for details.
}

\examples{
MantelHaenszel_estimate_stratified_2x2(doll_hill_1950)
MantelHaenszel_estimate_stratified_2x2(hine_1989)
}
