% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McNemar_exact_cond_test_paired_2x2.R
\name{McNemar_exact_cond_test_paired_2x2}
\alias{McNemar_exact_cond_test_paired_2x2}
\title{The McNemar exact conditional test}
\usage{
McNemar_exact_cond_test_paired_2x2(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (a 2x2 matrix)}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
The exact p-value based on the binomial distribution
}
\description{
The McNemar exact conditional test

Described in Chapter 8 "The Paired 2x2 Table"
}
\examples{
McNemar_exact_cond_test_paired_2x2(bentur_2009)
McNemar_exact_cond_test_paired_2x2(cavo_2012)
McNemar_exact_cond_test_paired_2x2(ezra_2010)
}
