% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MidP_multinomial_test_1xc.R
\name{MidP_multinomial_test_1xc}
\alias{MidP_multinomial_test_1xc}
\title{The mid-P multinomial test for multinomial probabilities}
\usage{
MidP_multinomial_test_1xc(n, pi0, printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (a 1xc vector, where c is the number of categories)}

\item{pi0}{given probabilities (a 1xc vector)}

\item{printresults}{display results (F = no, T = yes)}
}
\value{
probability value
}
\description{
The mid-P multinomial test for multinomial probabilities

Described in Chapter 3 "The 1xc Table and the Multinomial Distribution"
}
\examples{
# Genotype counts for SNP rs 6498169 in RA patients
\dontrun{
MidP_multinomial_test_1xc(n = snp6498169$complete$n, pi0 = snp6498169$complete$pi0)
}
# subset of 10 patients
MidP_multinomial_test_1xc(n = snp6498169$subset$n, pi0 = snp6498169$subset$pi0)
}
