% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mills_graubard_1987}
\alias{mills_graubard_1987}
\title{Alcohol consumption and malformations}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 2 columns.
}
\usage{
mills_graubard_1987
}
\description{
Alcohol consumption and malformations
}
\references{
Mills and Graubard (1987)
}
\keyword{datasets}
