% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ML_estimates}
\alias{ML_estimates}
\title{Calculate ML estimates}
\usage{
ML_estimates(...)
}
\arguments{
\item{...}{arguments passed to methods}
}
\description{
Calculate ML estimates
}
\note{
This function has little use to the user, it is exported for confirmity
to R package standards.
}
