% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{perondi_2004}
\alias{perondi_2004}
\title{An RCT of high vs standard dose of epinephrine}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
perondi_2004
}
\description{
An RCT of high vs standard dose of epinephrine
}
\references{
Perondi et al. (2004)
}
\keyword{datasets}
