% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{peterson_2007}
\alias{peterson_2007}
\title{Pretherapy susceptability of pathogens}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
peterson_2007
}
\description{
Pretherapy susceptability of pathogens
}
\references{
Peterson et al. (2007)
}
\keyword{datasets}
