% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peto_homogeneity_test_stratified_2x2.R
\name{Peto_homogeneity_test_stratified_2x2}
\alias{Peto_homogeneity_test_stratified_2x2}
\title{The Peto test for homogeneity of odds ratios over strata}
\usage{
Peto_homogeneity_test_stratified_2x2(n)
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}
}
\description{
The Peto test for homogeneity of odds ratios over strata

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\examples{
# Smoking and lung cancer (Doll and Hill, 1950)
Peto_homogeneity_test_stratified_2x2(doll_hill_1950)

# Prophylactice use of Lidocaine in myocardial infarction (Hine et al., 1989)
Peto_homogeneity_test_stratified_2x2(hine_1989)

}
