% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.contingencytables_result}
\alias{print.contingencytables_result}
\title{Output from a contingency tables method}
\usage{
\method{print}{contingencytables_result}(x, as_list = FALSE, ...)
}
\arguments{
\item{x}{The output from a function from the \link{contingencytables} package}

\item{as_list}{Print the elements of \code{x} as a list}

\item{...}{unused (kept for consistency with the generic \code{\link[base:print]{base::print()}})}
}
\description{
Output from a contingency tables method
}
