% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBG_test_and_CI_stratified_2x2.R
\name{RBG_test_and_CI_stratified_2x2}
\alias{RBG_test_and_CI_stratified_2x2}
\title{The RBG test and CI for a common odds ratio}
\usage{
RBG_test_and_CI_stratified_2x2(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}

\item{alpha}{the nominal level, e.g. 0.05 for 95# CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The RBG test and CI for a common odds ratio

(A Wald-type test and CI based on the Mantel-Haenszel estimate)

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\examples{
# Smoking and lung cancer (Doll and Hill, 1950)
RBG_test_and_CI_stratified_2x2(doll_hill_1950)

# Prophylactice use of Lidocaine in myocardial infarction (Hine et al., 1989)
RBG_test_and_CI_stratified_2x2(hine_1989)

}
