% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{Package script}
\usage{
reformat(filename, saveOutput = FALSE)
}
\arguments{
\item{filename}{name of the file}

\item{saveOutput}{if \code{TRUE}, original is printed and \code{filename} is overwritten. Defaults to \code{FALSE}, in which case reformatted file is printed}
}
\value{
text converted to R, printed to screen or replacing input file
}
\description{
Reformat script to fit package
}
\note{
This function is used to expedite conversion of the original scripts
into a package format
}
\author{
Waldir Leoncio
}
