% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reformatScript}
\alias{reformatScript}
\title{Package script}
\usage{
reformatScript(filename, saveOutput = FALSE)
}
\arguments{
\item{filename}{name of the file}

\item{saveOutput}{if \code{TRUE}, \code{filename} is overwritten. Defaults to \code{FALSE}}
}
\value{
text converted to R, printed to screen or replacing input file
}
\description{
Ad-hoc packages a function script
}
\note{
This function is used to expedite conversion of the original scripts
into a package format
}
\author{
Waldir Leoncio
}
