% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{reload_chapter}
\alias{reload_chapter}
\title{Reload chapter functions}
\usage{
reload_chapter(chap_num)
}
\arguments{
\item{chap_num}{Number of book chapter (from 2 to 10)}
}
\description{
Reloads functions from a chapter
}
\details{
This is a wrapper for \code{unload_chapter()} followed by
\code{load_chapter()}. It was made to ease the package development
process, but might be useful for end users who thinks something
might have gone wrong when loading/unloading a chapter and wishes
to reload it.
}
\author{
Waldir Leoncio
}
