% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spearman_correlation_coefficient_rxc.R
\name{Spearman_correlation_coefficient_rxc}
\alias{Spearman_correlation_coefficient_rxc}
\title{The Spearman correlation coefficient}
\usage{
Spearman_correlation_coefficient_rxc(n, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{alpha}{the nominal significance level, used to compute a 100(1-alpha)# confidence interval}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A list containing the Spearman correlation coefficient, and the Fieller and Bonett-Wright confidence intervals
}
\description{
The Spearman correlation coefficient

Described in Chapter 7 "The rxc Table"
}
\examples{
Spearman_correlation_coefficient_rxc(table_7.7)
\dontrun{
  Spearman_correlation_coefficient_rxc(table_7.8)
  Spearman_correlation_coefficient_rxc(table_7.9)
}
}
