% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuart_test_paired_cxc.R
\name{Stuart_test_paired_cxc}
\alias{Stuart_test_paired_cxc}
\title{The Stuart test for marginal homogeneity}
\usage{
Stuart_test_paired_cxc(n)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Stuart test for marginal homogeneity

Described in Chapter 9 "The Paired cxc Table"
}
\examples{
# Pretherapy susceptability of pathogens (Peterson et al., 2007)
Stuart_test_paired_cxc(peterson_2007)
}
