% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.3}
\alias{table_7.3}
\alias{vanbalen_2003}
\title{Treatment for ear infection}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 2 columns.
}
\usage{
table_7.3
vanbalen_2003
}
\description{
Status after 21 days treatment of the ear infection acute otitis
externa (Van Balen et al., 2003).

Van Balen et al. (2003) report a randomized, double-blind, controlled trial
comparing three treatments for an ear infection. The numbers and proportions
of patients reported cured and not cured after 21 days of treatment are
summarized in Table 7.3. Because there is no ordering between the treatments,
we regard Table 7.3 as an unordered 3 × 2 table.
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)

Van Balen et al. (2003)
}
\keyword{datasets}
