% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.4}
\alias{table_7.4}
\alias{mangerud_2014_PA}
\title{Psychiatric Diagnoses and Physical Activity}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 2 columns.
}
\usage{
table_7.4
mangerud_2014_PA
}
\description{
Psychiatric diagnoses and participation in team
sports (Mangerud et al., 2014)

Table 7.4 shows the number of subjects participating in team sports within
each of six psychiatric diagnoses, based on data from a study of physical
activity in adolescents aged 13 to 18 years who were referred to a child and
adolescent psychiatric clinic from 2009 to 2001 (Mangerud et al., 2014). The
psychiatric diagnoses are unordered, and we shall treat this as an unordered
6 x 2 table
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)
}
\keyword{datasets}
