% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.5}
\alias{table_7.5}
\title{Psychiatric diag. vs BMI with hyperkinetic disorders as reference category}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 3 columns.
}
\usage{
table_7.5
}
\description{
Psychiatric diag. vs BMI with hyperkinetic disorders as reference category
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)
}
\keyword{datasets}
