% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.6}
\alias{table_7.6}
\alias{lund_2012}
\title{Low Birth Weight vs psychiatric morbitidy with control as reference category}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
table_7.6
lund_2012
}
\description{
Categories of birth weight and psychiatric problems at age 20
years (Lund et al., 2012).

Lund et al. (2012) report psychiatric morbidity in young adulthood in two low
birth weight groups and a control group. The subjects were born between 1986
and 1988. The very low birth weight (VLBW) group consisted of babies born
preterm with birth weight up to 1500 grams. The small for gestational age at
term (SGA) group was born at term with birth weight below the 10th percentile
adjusted for gestational age, sex, and parity. The control group was born at
term, and was not small for gestational age. Table 7.6 shows the severity
level of psychiatric problems at age 20 years. We shall regard the birth
groups as unordered; however, the diagnostic groups are naturally ordered.
Hence, Table 7.6 is a singly ordered 3 × 3 table with unordered rows and
ordered columns.
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)

Lund et al. (2012)
}
\keyword{datasets}
