% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.8}
\alias{table_7.8}
\alias{bofin_2004}
\title{Breast Tumor}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 5 columns.
}
\usage{
table_7.8
bofin_2004
}
\description{
Nuclear pleomorphism from fine needle aspiration smears and
breast tumor type (Bofin et al., 2004).

Bofin et al. (2004) studied associations between different findings in fine
needle aspiration (FNA) smears from breast tumors and the final histological
diagnosis of tumor type in 133 patients. The aim of the study was to identify
variables developed from FNA smears that could differentiate between
the different tumor diagnoses. Table 7.8 presents the cross-classification of
the FNA variable nuclear pleomorphism with tumor types. Both variables can be
considered as ordered, with tumor type ordered from benign (as in NPBD) to
most malign (as in IDC).
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)

Bofin et al. (2004)
}
\keyword{datasets}
