% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.9}
\alias{table_7.9}
\alias{breidablik_2008}
\title{Self-rated health (Table 7.9)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
table_7.9
breidablik_2008
}
\description{
Self-rated health for 12 to 17 years old adolescents in
Young-HUNT 1 and four years later in Young-HUNT 2 (Breidablik et al., 2008).

In the HUNT study (Nord-Trøndelag county health survey), one of the questions
is: “How is your overall health at the moment?” The outcome categories
are “Very good”, “Good”, “Not very good”, and “Poor”. Table 7.9 shows
the counts for the adolescents aged 12 to 17 years in 1995 to 1997
(Young-HUNT 1), and for the same individuals four years later (Young-HUNT 2;
Breidablik et al. (2008)). Both the rows and the columns are ordered. In
this example, it may be appropriate to regard self-rated health as an
unobserved (latent) continuous variable, where only a categorized version has
been observed. Table 7.9 is actually an example of a paired c × c table with
ordinal data.
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)

Breidablik et al. (2008)
}
\keyword{datasets}
