% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tang_asymptotic_score_CI_paired_2x2.R
\name{Tang_asymptotic_score_CI_paired_2x2}
\alias{Tang_asymptotic_score_CI_paired_2x2}
\title{The Tang asymptotic score confidence interval for the ratio of paired probabilities}
\usage{
Tang_asymptotic_score_CI_paired_2x2(n, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A list containing lower, upper and point estimates of the statistic
}
\description{
The Tang asymptotic score confidence interval for the ratio of paired probabilities

Described in Chapter 8 "The Paired 2x2 Table"
}
\examples{
# Airway hyper-responsiveness before and after stem cell transplantation
# (Bentur et al., 2009)
n <- bentur_2009
Tang_asymptotic_score_CI_paired_2x2(n)

# Complete response before and after consolidation therapy
# (Cavo et al., 2012)
n <- cavo_2012
Tang_asymptotic_score_CI_paired_2x2(n)

}
