% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_1x2_table_tests.R
\name{the_1x2_table_tests}
\alias{the_1x2_table_tests}
\title{The 1x2 Table tests}
\usage{
the_1x2_table_tests(X, n, pi0)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{pi0}{a given probability}
}
\value{
NULL. This function should be called for its printed output
}
\description{
The 1x2 Table tests
}
\examples{
# Example: The number of 1st order male births (Singh et al. 2010)
the_1x2_table_tests(X = 250, n = 533, pi0 = 0.513)
# Example: The number of 2nd order male births (Singh et al. 2010)
the_1x2_table_tests(X = 204, n = 412, pi0 = 0.513)
# Example: The number of 3rd order male births (Singh et al. 2010)
the_1x2_table_tests(X = 103, n = 167, pi0 = 0.513)
# Example: The number of 4th order male births (Singh et al. 2010)
the_1x2_table_tests(X = 33, n = 45, pi0 = 0.513)
# Example: Ligarden et al. (2010)
the_1x2_table_tests(X = 13, n = 16, pi0 = 0.5)
}
