% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_1xc_table_CIs.R
\name{the_1xc_table_CIs}
\alias{the_1xc_table_CIs}
\title{The 1xc table CIs}
\usage{
the_1xc_table_CIs(n, alpha = 0.05)
}
\description{
The 1xc table CIs
}
\examples{
load_chapter(3)
# Genotype counts for SNP rs 6498169 in RA patients
the_1xc_table_CIs(n=c(276, 380, 118))
}
