% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_2x2_table_CIs_ratio.R
\name{the_2x2_table_CIs_ratio}
\alias{the_2x2_table_CIs_ratio}
\title{The 2x2 table CIs ratio}
\usage{
the_2x2_table_CIs_ratio(n, alpha = 0.05)
}
\arguments{
\item{n}{frequency matrix}

\item{alpha}{type I error}
}
\value{
NULL. This function should be called for its printed output
}
\description{
Wrapper for \verb{_CI_2x2} functions on Chapter 4.
}
\section{Table orientation}{
 In most study designs, rows designate a grouping of the data, for
instance, into treatment or exposure groups, while the columns indicate a
dichotomous outcome, usually with the first column representing the outcome
of interest. This designation of rows and columns may not be relevant in all
study designs, please see the introduction to chapter 4 for details.
}

\examples{
# An RCT of high vs standard dose of epinephrine (Perondi et al., 2004)
the_2x2_table_CIs_ratio(perondi_2004)

# The association between CHRNA4 genotype and XFS (Ritland et al., 2007)
the_2x2_table_CIs_ratio(ritland_2007)

}
\seealso{
the_2x2_table_CIs_difference the_2x2_table_CIs_OR the_2x2_table_tests
}
