% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_paired_2x2_table_CIs_ratio.R
\name{the_paired_2x2_table_CIs_ratio}
\alias{the_paired_2x2_table_CIs_ratio}
\title{The Paired 2x2 table CIs ratio}
\usage{
the_paired_2x2_table_CIs_ratio(n, alpha = 0.05)
}
\arguments{
\item{n}{frequency matrix}

\item{alpha}{type I error}
}
\description{
The Paired 2x2 table CIs ratio
}
\examples{
load_chapter(8)
# Airway hyper-responsiveness before and after stem cell transplantation
# (Bentur et al., 2009)
n <- rbind(c(1, 1), c(7, 12))
the_paired_2x2_table_CIs_ratio(n)

# Complete response before and after consolidation therapy
# (Cavo et al., 2012)
n <- rbind(c(59, 6), c(16, 80))
the_paired_2x2_table_CIs_ratio(n)

unload_chapter(8)
}
