% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_paired_cxc_table_ordinal.R
\name{the_paired_cxc_table_ordinal}
\alias{the_paired_cxc_table_ordinal}
\title{The Paired CxC table - ordinal}
\usage{
the_paired_cxc_table_ordinal(n, a, alpha = 0.05)
}
\arguments{
\item{n}{the total number of observations}

\item{a}{scores assigned to the outcome categories}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
NULL. This function should be called for its printed output.
}
\description{
The Paired CxC table - ordinal
}
\examples{
the_paired_cxc_table_ordinal(fischer_1999, c(8, 3.5, 0, -3.5, -8))
}
