% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_rx2_table.R
\name{the_rx2_table}
\alias{the_rx2_table}
\title{The rx2 table}
\usage{
the_rx2_table(n, alpha = 0.05, direction = "increasing", skip_exact = FALSE)
}
\arguments{
\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs#'}

\item{direction}{the direction of the success probabilities}

\item{skip_exact}{If \code{FALSE}, skips the exact conditional and mid-P tests
for unspecific ordering (often saves calculation time)
("increasing" or "decreasing")}
}
\description{
The rx2 table
}
\examples{
load_chapter(5)
n.0 <- rbind(c(48, 17066), c(38, 14464), c(5, 788), c(1, 126), c(1, 37))
n.1 <- rbind(c(8, 53), c(10, 48), c(11, 100), c(22, 102), c(6, 129))
a.0 <- c(1, 2, 3, 4, 5)
a.1 <- c(7, 22, 37, 48, 55)

# Alcohol consumption and malformations (Mills and Graubard, 1987)
direction <- 'increasing'
n <- n.0
a <- a.0
the_rx2_table(n, a, direction)

# Elevated troponin T levels in stroke patients (Indredavik et al., 2008)
direction <- 'decreasing'
n <- n.1
a <- a.0       # equally spaced scores
a <- a.1    # mid-interval scores
the_rx2_table(n, a, direction)

unload_chapter(5)
}
