% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{unload_chapter}
\alias{unload_chapter}
\title{Unload functions from a chapter}
\usage{
unload_chapter(chap_num)
}
\arguments{
\item{chap_num}{Number of book chapter (from 2 to 10)}
}
\value{
Unloads the functions from that chapter
}
\description{
Unloads all the functions associated with a specific book chapter.
}
\author{
Waldir Leoncio
}
