% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Wald_test_and_CI_marginal_mean_ranks_paired_cxc.R
\name{Wald_test_and_CI_marginal_mean_ranks_paired_cxc}
\alias{Wald_test_and_CI_marginal_mean_ranks_paired_cxc}
\title{The Wald test and confidence interval for the difference between
marginal mean ranks / ridits}
\usage{
Wald_test_and_CI_marginal_mean_ranks_paired_cxc(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Wald test and confidence interval for the difference between
marginal mean ranks / ridits

Described in Chapter 9 "The Paired cxc Table"
}
\examples{
# A comparison between serial and retrospective measurements
# (Fischer et al., 1999)
Wald_test_and_CI_marginal_mean_ranks_paired_cxc(fischer_1999)
}
