% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaptistaPike_exact_conditional_CI_2x2.R
\name{BaptistaPike_exact_conditional_CI_2x2}
\alias{BaptistaPike_exact_conditional_CI_2x2}
\title{The Baptista-Pike exact conditional confidence interval for the odds
ratio}
\usage{
BaptistaPike_exact_conditional_CI_2x2(n, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95# CIs}

\item{printresults}{display results (F = no, T = yes)}
}
\value{
A data frame containing lower, upper and point estimates of the
statistic
}
\description{
The Baptista-Pike exact conditional confidence interval for the
odds ratio

Described in Chapter 4 "The 2x2 Table"
}
\examples{
BaptistaPike_exact_conditional_CI_2x2(tea)
BaptistaPike_exact_conditional_CI_2x2(perondi_2004)
BaptistaPike_exact_conditional_CI_2x2(lampasona_2013)
BaptistaPike_exact_conditional_CI_2x2(ritland_2007)
}
