% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaptistaPike_midP_CI_2x2.R
\name{BaptistaPike_midP_CI_2x2}
\alias{BaptistaPike_midP_CI_2x2}
\title{The Baptista-Pike mid-P confidence interval for the odds ratio}
\usage{
BaptistaPike_midP_CI_2x2(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Baptista-Pike mid-P confidence interval for the odds ratio

Described in Chapter 4 "The 2x2 Table"
}
\examples{
BaptistaPike_midP_CI_2x2(tea)
BaptistaPike_midP_CI_2x2(perondi_2004)
BaptistaPike_midP_CI_2x2(lampasona_2013)
BaptistaPike_midP_CI_2x2(ritland_2007)
}
