% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blaker_exact_CI_1x2.R
\name{Blaker_exact_CI_1x2}
\alias{Blaker_exact_CI_1x2}
\title{The Blaker exact confidence interval}
\usage{
Blaker_exact_CI_1x2(X, n, alpha = 0.05)
}
\arguments{
\item{X}{the number of successes}

\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Blaker exact confidence interval for the binomial
probability
Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"
}
\examples{
Blaker_exact_CI_1x2(singh_2010["1st", "X"], singh_2010["1st", "n"])
Blaker_exact_CI_1x2(singh_2010["2nd", "X"], singh_2010["2nd", "n"])
Blaker_exact_CI_1x2(singh_2010["3rd", "X"], singh_2010["3rd", "n"])
with(singh_2010["4th", ], Blaker_exact_CI_1x2(X, n)) # alternative syntax
Blaker_exact_CI_1x2(ligarden_2010["X"], ligarden_2010["n"])
}
\references{
Blaker H (2000) Confidence curves and improved exact
confidence intervals for discrete distributions. The Canadian Journal of
Statistics; 28:783-798
}
