% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bonferroni_type_CIs_paired_cxc.R
\name{Bonferroni_type_CIs_paired_cxc}
\alias{Bonferroni_type_CIs_paired_cxc}
\title{Bonferroni-type confidence intervals for differences of marginal probabilities}
\usage{
Bonferroni_type_CIs_paired_cxc(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
Bonferroni-type confidence intervals for differences of marginal probabilities

Described in Chapter 9 "The Paired kxk Table"
}
\examples{
Bonferroni_type_CIs_paired_cxc(peterson_2007)
}
