% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap4}
\alias{chap4}
\title{Chapter 4: The 2x2 Table}
\description{
These are the functions related to chapter 4:
\enumerate{
\item{Adjusted_inv_sinh_CI_OR_2x2}
\item{Adjusted_inv_sinh_CI_ratio_2x2}
\item{Adjusted_log_CI_2x2}
\item{AgrestiCaffo_CI_2x2}
\item{Wald_CI_2x2}
\item{BaptistaPike_exact_conditional_CI_2x2}
\item{BaptistaPike_midP_CI_2x2}
\item{Cornfield_exact_conditional_CI_2x2}
\item{Cornfield_midP_CI_2x2}
\item{Fisher_exact_test_2x2}
\item{Exact_unconditional_test_2x2}
\item{Fisher_midP_test_2x2}
\item{Gart_adjusted_logit_CI_2x2}
\item{Independence_smoothed_logit_CI_2x2}
\item{Inv_sinh_CI_OR_2x2}
\item{Inv_sinh_CI_ratio_2x2}
\item{Katz_log_CI_2x2}
\item{Koopman_asymptotic_score_CI_2x2}
\item{LR_test_2x2}
\item{Mee_asymptotic_score_CI_2x2}
\item{MiettinenNurminen_asymptotic_score_CI_difference_2x2}
\item{MiettinenNurminen_asymptotic_score_CI_OR_2x2}
\item{MiettinenNurminen_asymptotic_score_CI_ratio_2x2}
\item{MOVER_R_Wilson_CI_OR_2x2}
\item{MOVER_R_Wilson_CI_ratio_2x2}
\item{Newcombe_hybrid_score_CI_2x2}
\item{Pearson_chi_squared_test_2x2}
\item{Pearson_chi_squared_test_CC_2x2}
\item{PriceBonett_approximate_Bayes_CI_2x2}
\item{Wald_CI_CC_2x2}
\item{Woolf_logit_CI_2x2}
\item{Uncorrected_asymptotic_score_CI_2x2}
\item{Z_unpooled_test_2x2}
\item{the_2x2_table_CIs_difference}
\item{the_2x2_table_CIs_OR}
\item{the_2x2_table_CIs_ratio}
\item{the_2x2_table_tests}
}
}
\note{
You can also print the list above with \code{list_functions(4)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
