% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters.R
\name{chap8}
\alias{chap8}
\title{Chapter 8: The Paired 2x2 Table}
\description{
These are the functions related to chapter 8:
\enumerate{
\item{BonettPrice_hybrid_Wilson_score_CI_CC_paired_2x2}
\item{BonettPrice_hybrid_Wilson_score_CI_paired_2x2}
\item{ClopperPearson_exact_CI_1x2_beta_version}
\item{McNemar_asymptotic_test_CC_paired_2x2}
\item{McNemar_asymptotic_test_paired_2x2}
\item{McNemar_exact_cond_test_paired_2x2}
\item{McNemar_exact_unconditional_test_paired_2x2}
\item{McNemar_midP_test_paired_2x2}
\item{Tang_asymptotic_score_CI_paired_2x2}
\item{Tango_asymptotic_score_CI_paired_2x2}
\item{MOVER_Wilson_score_CI_paired_2x2}
\item{Newcombe_square_and_add_CI_paired_2x2}
\item{Transformed_Blaker_exact_CI_paired_2x2}
\item{Transformed_Clopper_Pearson_exact_CI_paired_2x2}
\item{Transformed_Clopper_Pearson_midP_CI_paired_2x2}
\item{Transformed_Wilson_score_CI_paired_2x2}
\item{Wald_CI_diff_paired_2x2}
\item{Wald_CI_diff_CC_paired_2x2}
\item{Wald_CI_AgrestiMin_paired_2x2}
\item{Wald_CI_BonettPrice_paired_2x2}
\item{Wald_CI_OR_Laplace_paired_2x2}
\item{Wald_CI_OR_paired_2x2}
\item{Wald_CI_ratio_paired_2x2}
\item{the_paired_2x2_table_CIs_difference}
\item{the_paired_2x2_table_CIs_OR}
\item{the_paired_2x2_table_CIs_ratio}
\item{the_paired_2x2_table_tests}
}
}
\note{
You can also print the list above with \code{list_functions(8)}.
}
\references{
\itemize{
\item Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of
Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL
\item https://contingencytables.com/
\item https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172
}
}
