% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranArmitage_CI_rx2.R
\name{CochranArmitage_CI_rx2}
\alias{CochranArmitage_CI_rx2}
\title{The Cochran-Armitage confidence interval for trend in the linear model}
\usage{
CochranArmitage_CI_rx2(n, a = seq_len(nrow(n)), alpha = 0.05)
}
\arguments{
\item{n}{the observed counts (an rx2 matrix)}

\item{a}{scores assigned to the rows}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Cochran-Armitage confidence interval for trend in the linear model

Described in Chapter 5 "The Ordered rx2 Table"
}
\examples{
CochranArmitage_CI_rx2(mills_graubard_1987, c(1, 2, 3, 4, 5))
CochranArmitage_CI_rx2(indredavik_2008, c(1, 2, 3, 4, 5))
}
