% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranMantelHaenszel_test_stratified_2x2.R
\name{CochranMantelHaenszel_test_stratified_2x2}
\alias{CochranMantelHaenszel_test_stratified_2x2}
\title{The Cochran-Mantel-Haenszel test of a common odds ratio}
\usage{
CochranMantelHaenszel_test_stratified_2x2(n)
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Cochran-Mantel-Haenszel test of a common odds ratio

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\section{Table orientation}{
 In most study designs, rows designate a grouping of the data, for
instance, into treatment or exposure groups, while the columns indicate a
dichotomous outcome, usually with the first column representing the outcome
of interest. This designation of rows and columns may not be relevant in all
study designs, please see the introduction to chapter 10 for details.
}

\examples{
CochranMantelHaenszel_test_stratified_2x2(doll_hill_1950)
CochranMantelHaenszel_test_stratified_2x2(hine_1989)
}
