% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulative_models_for_rxc.R
\name{Cumulative_models_for_rxc}
\alias{Cumulative_models_for_rxc}
\title{Cumulative logit and probit models}
\usage{
Cumulative_models_for_rxc(n, linkfunction = "logit", alpha = 0.05)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{linkfunction}{either "logit" or "probit"}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
Cumulative logit and probit models

Described in Chapter 7 "The rxc Table"
}
\examples{
Cumulative_models_for_rxc(table_7.5)
Cumulative_models_for_rxc(table_7.6)
}
