% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.onAttach}
\alias{.onAttach}
\title{Prints welcome message on package load}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library location. See \code{?base::.onAttach} for details}

\item{pkgname}{package name. See \code{?base::.onAttach} for details}
}
\description{
Prints package version number and welcome message on package load
}
