% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ear_infection}
\alias{ear_infection}
\title{Treatment for ear infection}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 2 columns.
}
\usage{
ear_infection
}
\description{
Treatment for ear infection
}
\keyword{datasets}
