% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exact_cond_midP_linear_rank_tests_2xc.R
\name{Exact_cond_midP_linear_rank_tests_2xc}
\alias{Exact_cond_midP_linear_rank_tests_2xc}
\title{Exact conditional and mid-P linear rank tests}
\usage{
Exact_cond_midP_linear_rank_tests_2xc(n, b = 0)
}
\arguments{
\item{n}{the observed table (a 2xc matrix)}

\item{b}{scores assigned to the columns (if b=0, midranks will be used as scores)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
Exact conditional and mid-P linear rank tests

Described in Chapter 6 "The Ordered 2xc Table"
}
\examples{
Exact_cond_midP_linear_rank_tests_2xc(lydersen_2012a)
Exact_cond_midP_linear_rank_tests_2xc(fontanella_2008)
}
