% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exact_cond_midP_tests_rxc.R
\name{Exact_cond_midP_tests_rxc}
\alias{Exact_cond_midP_tests_rxc}
\title{Exact conditional and mid-P tests for the rxc table}
\usage{
Exact_cond_midP_tests_rxc(n)
}
\arguments{
\item{n}{the observed counts (an rxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
Exact conditional and mid-P tests for the rxc table:
the Fisher-Freeman-Halton, Pearson, likelihood ratio, Kruskal-Wallis,
linear-by-linear, and Jonckheere-Terpstra tests.

Described in Chapter 7 "The rxc Table"
}
\note{
Works only for 3x2 and 3x3 tables
}
\examples{
Exact_cond_midP_tests_rxc(table_7.3)  # a 3x2 table
\dontrun{
  Exact_cond_midP_tests_rxc(table_7.6) # a 3x3 table
}
}
