% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_exact_test_2x2.R
\name{Fisher_exact_test_2x2}
\alias{Fisher_exact_test_2x2}
\title{The Fisher exact test for association in 2x2 tables}
\usage{
Fisher_exact_test_2x2(n, statistic = "Pearson", printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{statistic}{'hypergeometric' (i.e. Fisher-Irwin; default), 'Pearson', or 'LR' (likelihood ratio)}

\item{printresults}{display results (F = no, T = yes)}
}
\value{
probability value
}
\description{
The Fisher exact test for association in 2x2 tables

Described in Chapter 4 "The 2x2 Table"
}
\examples{
n <- tea # Example: A lady tasting a cup of tea
Fisher_exact_test_2x2(n)
n <- perondi_2004 # Example: Perondi et al. (2004)
Fisher_exact_test_2x2(n)
n <- lampasona_2013 # Example: Lampasona et al. (2013)
Fisher_exact_test_2x2(n)
n <- ritland_2007 # Example: Ritland et al. (2007)
Fisher_exact_test_2x2(n)
}
