% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_exact_test_2x2.R
\name{Fisher_exact_test_2x2}
\alias{Fisher_exact_test_2x2}
\title{The Fisher exact test for association in 2x2 tables}
\usage{
Fisher_exact_test_2x2(n, statistic = "Pearson")
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{statistic}{'hypergeometric' (i.e. Fisher-Irwin; default), 'Pearson', or 'LR' (likelihood ratio)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Fisher exact test for association in 2x2 tables

Described in Chapter 4 "The 2x2 Table"
}
\examples{
Fisher_exact_test_2x2(tea)
Fisher_exact_test_2x2(perondi_2004)
Fisher_exact_test_2x2(lampasona_2013)
Fisher_exact_test_2x2(ritland_2007)
}
