% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherFreemanHalton_asymptotic_test_rxc.R
\name{FisherFreemanHalton_asymptotic_test_rxc}
\alias{FisherFreemanHalton_asymptotic_test_rxc}
\title{The Fisher-Freeman-Halton asymptotic test for unordered rxc tables}
\usage{
FisherFreemanHalton_asymptotic_test_rxc(n)
}
\arguments{
\item{n}{the observed counts (an rxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Fisher-Freeman-Halton asymptotic test for unordered rxc tables

Described in Chapter 7 "The rxc Table"
}
\note{
May not give results for all tables, due to overflow
}
\examples{
FisherFreemanHalton_asymptotic_test_rxc(table_7.3)
}
