% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleissEveritt_test_paired_cxc.R
\name{FleissEveritt_test_paired_cxc}
\alias{FleissEveritt_test_paired_cxc}
\title{The Fleiss-Everitt version of the Stuart test for marginal homogeneity}
\usage{
FleissEveritt_test_paired_cxc(n)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Fleiss-Everitt version of the Stuart test for marginal homogeneity

Described in Chapter 9 "The Paired cxc Table"
}
\examples{
FleissEveritt_test_paired_cxc(fleiss_2003)
}
