% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleissLevinPaik_test_paired_cxc.R
\name{FleissLevinPaik_test_paired_cxc}
\alias{FleissLevinPaik_test_paired_cxc}
\title{The Fleiss-Levin-Paik test for three-level ordinal outcomes}
\usage{
FleissLevinPaik_test_paired_cxc(n)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Fleiss-Levin-Paik test for three-level ordinal outcomes

Described in Chapter 9 "The Paired cxc Table"
}
\examples{
# Pretherapy susceptability of pathogens *without the N / A category*
FleissLevinPaik_test_paired_cxc(peterson_2007[-4, -4])
}
