% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_coefficient_rxc.R
\name{gamma_coefficient_rxc}
\alias{gamma_coefficient_rxc}
\title{The gamma coefficient}
\usage{
gamma_coefficient_rxc(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
a list containing the gamma coefficient, the number of concordant
pairs and the number of discordant pairs
}
\description{
The gamma coefficient

Described in Chapter 7 "The rxc Table"
}
\examples{
\dontrun{
  gamma_coefficient_rxc(table_7.7)
  gamma_coefficient_rxc(table_7.8)
  gamma_coefficient_rxc(table_7.9)
}
}
