% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_coefficient_rxc.R
\name{gamma_coefficient_rxc}
\alias{gamma_coefficient_rxc}
\title{The gamma coefficient}
\usage{
gamma_coefficient_rxc(n)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The gamma coefficient

Described in Chapter 7 "The rxc Table"
}
\examples{
gamma_coefficient_rxc(table_7.7)
gamma_coefficient_rxc(table_7.8)
gamma_coefficient_rxc(table_7.9)
}
