% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_sinh_CI_OR_2x2.R
\name{Inv_sinh_CI_OR_2x2}
\alias{Inv_sinh_CI_OR_2x2}
\title{The inverse hyperbolic sine confidence interval for the odds ratio}
\usage{
Inv_sinh_CI_OR_2x2(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
A data frame containing lower, upper and point estimates of the statistic
}
\description{
The inverse hyperbolic sine confidence interval for the odds ratio

Described in Chapter 4 "The 2x2 Table"
}
\examples{
Inv_sinh_CI_OR_2x2(lampasona_2013)
Inv_sinh_CI_OR_2x2(ritland_2007)
}
