% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_sinh_CI_ratio_2x2.R
\name{Inv_sinh_CI_ratio_2x2}
\alias{Inv_sinh_CI_ratio_2x2}
\title{The inverse hyperbolic sine confidence interval for the ratio of probabilities}
\usage{
Inv_sinh_CI_ratio_2x2(n, alpha = 0.05, printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A data frame containing lower, upper and point estimates of the statistic
}
\description{
The inverse hyperbolic sine confidence interval for the ratio of probabilities

Described in Chapter 4 "The 2x2 Table"
}
\examples{
Inv_sinh_CI_ratio_2x2(perondi_2004)
Inv_sinh_CI_ratio_2x2(ritland_2007)
}
