% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kendalls_tau_b_rxc.R
\name{Kendalls_tau_b_rxc}
\alias{Kendalls_tau_b_rxc}
\title{Kendall's tau-b with confidence interval based on the Fieller standard deviation}
\usage{
Kendalls_tau_b_rxc(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{alpha}{the nominal significance level, used to compute a 100(1-alpha)\% confidence interval}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
Kendall's tau-b with confidence interval based on the Fieller standard deviation

Described in Chapter 7 "The rxc Table"
}
\examples{
Kendalls_tau_b_rxc(table_7.7)
Kendalls_tau_b_rxc(table_7.8)
Kendalls_tau_b_rxc(table_7.9)
}
