% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KruskalWallis_asymptotic_test_rxc.R
\name{KruskalWallis_asymptotic_test_rxc}
\alias{KruskalWallis_asymptotic_test_rxc}
\title{The Kruskal-Wallis asymptotic test for singly ordered rxc tables}
\usage{
KruskalWallis_asymptotic_test_rxc(n)
}
\arguments{
\item{n}{the observed counts (an rxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Kruskal-Wallis asymptotic test for singly ordered rxc tables

Described in Chapter 7 "The rxc Table"
}
\examples{
KruskalWallis_asymptotic_test_rxc(table_7.5)
KruskalWallis_asymptotic_test_rxc(table_7.6)
}
