% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_by_linear_test_rxc.R
\name{linear_by_linear_test_rxc}
\alias{linear_by_linear_test_rxc}
\title{The linear-by-linear test for association}
\usage{
linear_by_linear_test_rxc(n, a = seq_len(ncol(n)), b = seq_len(nrow(n)))
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{a}{scores assigned to the rows}

\item{b}{scores assigned to the columns}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The linear-by-linear test for association

Described in Chapter 7 "The rxc Table"
}
\examples{
linear_by_linear_test_rxc(table_7.7)
linear_by_linear_test_rxc(table_7.8)
linear_by_linear_test_rxc(table_7.9)
}
