% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test_2x2.R
\name{LR_test_2x2}
\alias{LR_test_2x2}
\title{The likelihood ratio test for association in 2x2 tables}
\usage{
LR_test_2x2(n)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The likelihood ratio test for association in 2x2 tables

Described in Chapter 4 "The 2x2 Table"
}
\examples{
LR_test_2x2(tea)
LR_test_2x2(perondi_2004)
LR_test_2x2(lampasona_2013)
LR_test_2x2(ritland_2007)
}
