% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MantelHaenszel_estimate_stratified_2x2.R
\name{MantelHaenszel_estimate_stratified_2x2}
\alias{MantelHaenszel_estimate_stratified_2x2}
\title{The Mantel-Haenszel estimate of the overall effect across strata}
\usage{
MantelHaenszel_estimate_stratified_2x2(n, link = "logit", printresults = TRUE)
}
\arguments{
\item{n}{the observed table (a 2x2xk matrix, where k is the number of strata)}

\item{link}{the link function ('linear', 'log', or 'logit')}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
a list respectively containing the inverse variance estimate of the overall effect (\code{estimate}), the stratum-specific effect estimates (\code{psihat}) and the weights (\code{w}).
}
\description{
The Mantel-Haenszel estimate of the overall effect across strata

Described in Chapter 10 "Stratified 2x2 Tables and Meta-Analysis"
}
\examples{
MantelHaenszel_estimate_stratified_2x2(doll_hill_1950)
MantelHaenszel_estimate_stratified_2x2(hine_1989)
}
