% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MantelHaenszel_test_2xc.R
\name{MantelHaenszel_test_2xc}
\alias{MantelHaenszel_test_2xc}
\title{The Mantel-Haenszel test of association with column scores}
\usage{
MantelHaenszel_test_2xc(n, b = 0)
}
\arguments{
\item{n}{the observed counts (a 2xc matrix)}

\item{b}{scores assigned to the columns (if b=0, midranks will be used as scores)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Mantel-Haenszel test of association with column scores

Described in Chapter 6 "The Ordered 2xc Table"
}
\examples{
MantelHaenszel_test_2xc(lydersen_2012a)
}
