% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McNemar_asymptotic_test_CC_paired_2x2.R
\name{McNemar_asymptotic_test_CC_paired_2x2}
\alias{McNemar_asymptotic_test_CC_paired_2x2}
\title{The McNemar asymptotic test with continuity correction}
\usage{
McNemar_asymptotic_test_CC_paired_2x2(n)
}
\arguments{
\item{n}{the observed table (a 2x2 matrix)}
}
\value{
The McNemar test statistic with continuity correction (\code{Z}) and the reference distribution (standard normal, \code{P})
}
\description{
The McNemar asymptotic test with continuity correction

Described in Chapter 8 "The Paired 2x2 Table"
}
\examples{
McNemar_asymptotic_test_CC_paired_2x2(bentur_2009)
McNemar_asymptotic_test_CC_paired_2x2(cavo_2012)
McNemar_asymptotic_test_CC_paired_2x2(ezra_2010)

}
