% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McNemar_asymptotic_test_paired_2x2.R
\name{McNemar_asymptotic_test_paired_2x2}
\alias{McNemar_asymptotic_test_paired_2x2}
\title{The McNemar asymptotic test}
\usage{
McNemar_asymptotic_test_paired_2x2(n)
}
\arguments{
\item{n}{the observed table (a 2x2 matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The McNemar asymptotic test

Described in Chapter 8 "The Paired 2x2 Table"
}
\examples{
McNemar_asymptotic_test_paired_2x2(bentur_2009)
McNemar_asymptotic_test_paired_2x2(cavo_2012)
McNemar_asymptotic_test_paired_2x2(ezra_2010)
}
