% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/MiettinenNurminen_asymptotic_score_CI_difference_2x2.R
\name{MiettinenNurminen_asymptotic_score_CI_difference_2x2}
\alias{MiettinenNurminen_asymptotic_score_CI_difference_2x2}
\title{The Mietinen-Nurminen asymptotic score confidence interval for the}
\usage{
MiettinenNurminen_asymptotic_score_CI_difference_2x2(
  n,
  alpha = 0.05,
  printresults = TRUE
)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A data frame containing lower, upper and point estimates of the statistic
}
\description{
The Mietinen-Nurminen asymptotic score confidence interval for the

difference between probabilities

Described in Chapter 4 "The 2x2 Table"
}
\examples{
# An RCT of high vs standard dose of epinephrine (Perondi et al., 2004):
n <- perondi_2004
MiettinenNurminen_asymptotic_score_CI_difference_2x2(n)
# The association between CHRNA4 genotype and XFS (Ritland et al., 2007):
n <- ritland_2007
MiettinenNurminen_asymptotic_score_CI_difference_2x2(n)
}
