% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOVER_Wilson_score_CI_paired_2x2.R
\name{MOVER_Wilson_score_CI_paired_2x2}
\alias{MOVER_Wilson_score_CI_paired_2x2}
\title{The MOVER Wilson score confidence interval for the ratio of paired probabilities}
\usage{
MOVER_Wilson_score_CI_paired_2x2(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The MOVER Wilson score confidence interval for the ratio of paired probabilities

Described in Chapter 8 "The Paired 2x2 Table"
}
\examples{
MOVER_Wilson_score_CI_paired_2x2(bentur_2009)
MOVER_Wilson_score_CI_paired_2x2(cavo_2012)

}
