% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearson_chi_squared_test_2x2.R
\name{Pearson_chi_squared_test_2x2}
\alias{Pearson_chi_squared_test_2x2}
\title{The Pearson chi-squared test for association in 2x2 tables}
\usage{
Pearson_chi_squared_test_2x2(n, printresults = TRUE)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{printresults}{display results (FALSE = no, TRUE = yes)}
}
\value{
A data frame containing the two-sided p-value, the statistic and the degrees of freedom
}
\description{
The Pearson chi-squared test for association in 2x2 tables

Described in Chapter 4 "The 2x2 Table"
}
\examples{
# Example: A lady tasting a cup of tea
n <- tea
Pearson_chi_squared_test_2x2(n)

# Example: Perondi et al. (2004)
n <- perondi_2004
Pearson_chi_squared_test_2x2(n)

# Example: Lampasona et al. (2013)
n <- lampasona_2013
Pearson_chi_squared_test_2x2(n)

# Example: Ritland et al. (2007)
n <- ritland_2007
Pearson_chi_squared_test_2x2(n)

}
