% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearson_residuals_rxc.R
\name{Pearson_residuals_rxc}
\alias{Pearson_residuals_rxc}
\title{The Pearson residuals and the standardized Pearson residuals}
\usage{
Pearson_residuals_rxc(n)
}
\arguments{
\item{n}{the observed counts (an rxc matrix)}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Pearson residuals and the standardized Pearson residuals

Described in Chapter 7 "The rxc Table"
}
\examples{
## Treatment for ear infection (van Balen et al., 2003):
Pearson_residuals_rxc(table_7.3)

## Psychiatric diagnoses vs PA (Mangerud et al., 2004):
Pearson_residuals_rxc(table_7.4)

## Psychiatric diag. vs BMI (Mangerud et al., 2004):
Pearson_residuals_rxc(table_7.5)
}
