% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheffe_type_CIs_paired_cxc.R
\name{Scheffe_type_CIs_paired_cxc}
\alias{Scheffe_type_CIs_paired_cxc}
\title{ScheffE-type confidence intervals for differences of marginal probabilities}
\usage{
Scheffe_type_CIs_paired_cxc(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
ScheffE-type confidence intervals for differences of marginal probabilities

Described in Chapter 9 "The Paired kxk Table"
}
\examples{
# Pretherapy susceptability of pathogens (Peterson et al., 2007)
Scheffe_type_CIs_paired_cxc(peterson_2007)
}
