% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheffe_type_CIs_rxc.R
\name{Scheffe_type_CIs_rxc}
\alias{Scheffe_type_CIs_rxc}
\title{The ScheffE-type simultaneous confidence intervals for the differences pi_1|i - pi_1|j}
\usage{
Scheffe_type_CIs_rxc(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed counts (an rx2 vector)}

\item{alpha}{the nominal level, e.g. 0.05 for 95# CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The ScheffE-type simultaneous confidence intervals for the differences pi_1|i - pi_1|j

Described in Chapter 7 "The rxc Table"
}
\examples{
# Example: Treatment for ear infection
Scheffe_type_CIs_rxc(table_7.3)
}
