% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Score_test_and_CI_marginal_mean_scores_paired_cxc.R
\name{Score_test_and_CI_marginal_mean_scores_paired_cxc}
\alias{Score_test_and_CI_marginal_mean_scores_paired_cxc}
\title{Score test and CI marginal mean scores paired CxC}
\usage{
Score_test_and_CI_marginal_mean_scores_paired_cxc(n, a, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (a cxc matrix)}

\item{a}{scores assigned to the outcome categories}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The score test and confidence interval for the difference
between marginal mean scores Described in Chapter 9 "The Paired cxc Table"
}
\examples{
# A comparison between serial and retrospective measurements
# (Fischer et al., 1999)
a <- c(8, 3.5, 0, -3.5, -8)
Score_test_and_CI_marginal_mean_scores_paired_cxc(fischer_1999, a)
}
