% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spearman_correlation_coefficient_rxc.R
\name{Spearman_correlation_coefficient_rxc}
\alias{Spearman_correlation_coefficient_rxc}
\title{The Spearman correlation coefficient}
\usage{
Spearman_correlation_coefficient_rxc(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed table (an rxc matrix)}

\item{alpha}{the nominal significance level, used to compute a 100(1-alpha)# confidence interval}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The Spearman correlation coefficient

Described in Chapter 7 "The rxc Table"
}
\examples{
Spearman_correlation_coefficient_rxc(table_7.7)
Spearman_correlation_coefficient_rxc(table_7.8)
Spearman_correlation_coefficient_rxc(table_7.9)
}
