% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_7.5}
\alias{table_7.5}
\alias{mangerud_2014_BMI}
\title{Psychiatric diag. vs BMI with hyperkinetic disorders as reference category}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 3 columns.
}
\usage{
table_7.5
mangerud_2014_BMI
}
\description{
Psychiatric diagnoses and weight categories based on age- and
sex-adjusted BMI (Mangerud et al., 2014).

Table 7.5 shows the number of thin, normal weight, and overweight subjects
within each of six psychiatric diagnoses, based on the same study as in
Section 7.2.2 (Mangerud et al., 2014). Body mass index (BMI) is calculated
as the weight in kg divided by the squared height in meters. In subjects
aged 18 years or older, the cut-off points for being categorized as thin,
normal weight, and overweight are BMI less than 18.5, BMI between 18.5 and
25, and BMI above 25, respectively. For younger subjects (below 18 years of
age), the categorization was done following internationally adopted cut-off
points for age and sex (Cole et al., 2000, 2007). For example, the cut-off
point for being overweight at age 13 is 21.91 for males and 22.58 for
females.
}
\references{
Fagerland MW, Lydersen S, Laake P (2017)

Mangerud et al. (2014)
}
\keyword{datasets}
