% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_1xc_table_CIs.R
\name{the_1xc_table_CIs}
\alias{the_1xc_table_CIs}
\title{The 1xc table CIs}
\usage{
the_1xc_table_CIs(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed counts (a 1xc vector, where c is the number of categories)}

\item{alpha}{the nominal level, e.g. 0.05 for 95# CIs}
}
\value{
A data frame containing lower, upper and point estimates of the statistic
}
\description{
The 1xc table CIs
}
\examples{
# Genotype counts for SNP rs 6498169 in RA patients
the_1xc_table_CIs(n = c(276, 380, 118))
}
