% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_2x2_table_CIs_OR.R
\name{the_2x2_table_CIs_OR}
\alias{the_2x2_table_CIs_OR}
\title{The 2x2 table CIs odds ratio}
\usage{
the_2x2_table_CIs_OR(n, alpha = 0.05)
}
\arguments{
\item{n}{frequency matrix}

\item{alpha}{type I error}
}
\value{
NULL. This function should be called for its printed output
}
\description{
Wrapper for \verb{_CI_OR_2x2} functions on Chapter 4.
}
\examples{
# Example: A lady tasting a cup of tea
the_2x2_table_CIs_OR(tea)

# Example: Perondi et al. (2004)
the_2x2_table_CIs_OR(perondi_2004)

# Example: Lampasona et al. (2013)
the_2x2_table_CIs_OR(lampasona_2013)

# Example: Ritland et al. (2007)
the_2x2_table_CIs_OR(ritland_2007)

}
