% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_paired_2x2_table_CIs_OR.R
\name{the_paired_2x2_table_CIs_OR}
\alias{the_paired_2x2_table_CIs_OR}
\title{The Paired 2x2 table CIs OR}
\usage{
the_paired_2x2_table_CIs_OR(n, alpha = 0.05)
}
\arguments{
\item{n}{frequency matrix}

\item{alpha}{type I error}
}
\value{
NULL. This function should be called for its printed output.
}
\description{
The Paired 2x2 table CIs OR
}
\section{Table orientation}{
 In most study designs, rows designate a grouping of the data, for
instance, into treatment or exposure groups, while the columns indicate a
dichotomous outcome, usually with the first column representing the outcome
of interest. This designation of rows and columns may not be relevant in all
study designs, please see the introduction to chapter 4 for details.
}

\examples{
the_paired_2x2_table_CIs_OR(ezra_2010)
}
