% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_paired_2x2_table_CIs_OR.R
\name{the_paired_2x2_table_CIs_OR}
\alias{the_paired_2x2_table_CIs_OR}
\title{The Paired 2x2 table CIs OR}
\usage{
the_paired_2x2_table_CIs_OR(n, alpha = 0.05)
}
\arguments{
\item{n}{frequency matrix}

\item{alpha}{type I error}
}
\description{
The Paired 2x2 table CIs OR
}
\examples{

# Floppy eyelid syndrome vs obstructive sleep apnea (Ezra et al., 2010)
n <- rbind(c(7, 25), c(2, 68))
the_paired_2x2_table_CIs_OR(n)


}
