% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_rx2_table.R
\name{the_rx2_table}
\alias{the_rx2_table}
\title{The rx2 table}
\usage{
the_rx2_table(n, alpha = 0.05, direction = "increasing", skip_exact = FALSE)
}
\arguments{
\item{n}{the total number of observations}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}

\item{direction}{the direction of the success probabilities}

\item{skip_exact}{If \code{FALSE}, skips the exact conditional and mid-P tests
for unspecific ordering (often saves calculation time)
("increasing" or "decreasing")}
}
\value{
NULL. This function should be called for its printed output.
}
\description{
The rx2 table
}
\examples{
the_rx2_table(mills_graubard_1987, skip_exact = TRUE)
the_rx2_table(indredavik_2008, direction = "decreasing", skip_exact = TRUE)
}
