% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncorrected_asymptotic_score_CI_2x2.R
\name{Uncorrected_asymptotic_score_CI_2x2}
\alias{Uncorrected_asymptotic_score_CI_2x2}
\title{The uncorrected asymptotic score confidence interval for the odds
ratio}
\usage{
Uncorrected_asymptotic_score_CI_2x2(n, alpha = 0.05)
}
\arguments{
\item{n}{the observed counts (a 2x2 matrix)}

\item{alpha}{the nominal level, e.g. 0.05 for 95\% CIs}
}
\value{
An object of the \link{contingencytables_result} class,
basically a subclass of \code{\link[base:list]{base::list()}}. Use the \code{\link[utils:str]{utils::str()}} function
to see the specific elements returned.
}
\description{
The uncorrected asymptotic score confidence interval for the
odds ratio

Described in Chapter 4 "The 2x2 Table"
}
\examples{
# A case-control study of GADA exposure on IPEX syndrome
# (Lampasona et al., 2013):
Uncorrected_asymptotic_score_CI_2x2(lampasona_2013)

# The association between CHRNA4 genotype and XFS (Ritland et al., 2007):
Uncorrected_asymptotic_score_CI_2x2(ritland_2007)

}
